refactor rename prms canopy module
